+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The changelog for **SwifterSwift**. Also see the [releases](https://github.com/S
 
 ### Added
 
+- **CGSize**:
+  - Added `aspectRatio` and `max` properties. [#662](https://github.com/SwifterSwift/SwifterSwift/pull/662) by [vyax](https://github.com/vyax)
+
 ### Changed
 
 ### Deprecated